requirements.txt
pandas
numpy
matplotlib
tensorflow
scikit-learn
python-binance
